Moved ClaimbuildRepository to correct package
