+++ src/main/java/com/ardaslegends/repository/claimbuild/ClaimbuildRepository.java
@@ -1,14 +1,13 @@
-package com.ardaslegends.repository;
+package com.ardaslegends.repository.claimbuild;
 
 import com.ardaslegends.domain.ClaimBuild;
-import com.ardaslegends.repository.claimbuild.ClaimbuildRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface ClaimBuildRepository extends JpaRepository<ClaimBuild, Long>, ClaimbuildRepositoryCustom {
+public interface ClaimbuildRepository extends JpaRepository<ClaimBuild, Long>, ClaimbuildRepositoryCustom {
 
     Optional<ClaimBuild> findClaimBuildByName(String name);
 }
